trendSIG is now functional and ready for tagging
trendSIG is a simple signal that works by checking all the trends given to it as a parameter.
- It does not check the slope strength each signal
- it also does not check the if a signal is above or below the other signal
- it relies on the summation of the general direction of all trend signals.
  if all signals show buy then it is a buy. conversely if all the signals show sell then it is a sell

Improvements: Another signal based on trendSIG which takes into account the slope as well as the actual position of the
each trend may be considered at a later stage. The slopeVarSIG checks only for slopes of signals. If the slope of the fastest
signal has to be the greatest. It is a great candidate for trendslopes as well.